ai-service/requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
deepface==0.0.99
opencv-python==4.13.0.92
numpy
faiss-cpu
requests
python-multipart
cloudinary
pillow
tf-keras
tensorflow